Factor out select module
